+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.1.0
+* Added news reducer and logic
+
+## 1.0.4
+* Fixed ticket pool id no being saved after selecting day in calendar.
+
+## 1.0.3
+* Fixed `clearCartLogic` in `cart` duck.
+
+## 1.0.2
+* Fixed `getInitialActiveStep` logic in TicketModalController.
+
+## 1.0.1
+* Fixed date parsing in `TicketModalControler` and `pool-parser` utils
+
 ## 1.0.0
 * Refactored cart